requirements.txt
gradio>=4.0.0
google-genai>=1.0.0
opencv-python>=4.8.0
moviepy<2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
